Following the first 3 iterations
First i made the initial source that is index.html to which i adedthe main heading. Then i created the recipe page by creating a new html file in the recipes directory which i created in the recipes of odin directory. i then linked the recipe page to the main index page. i then made a new directory for the images to which i added all imagec i will use. I then followed the steps of iteration 3 and made an unordered list and a ordered list.