Making another recipe page and correcting errors
First i corrected the description lines on peanut_butter_spider_cookies, and added the paragraph to the links on the index page. Then i made another recipe file in the recipes directory to which i added an image and one unordered and one ordered list. Lastly i linked this file to the main index page.